feat : book rating 관련 로직 & jsp 싹다 수정.

+++ src/main/java/com/book4w/book4w/dto/response/BoardListResponseDTO.java
@@ -14,7 +14,7 @@ public class BoardListResponseDTO {
     private final String pub;
     private final int year;
     private final String coverImage;
-    private final int rating;
+    private final double rating;
     private final int reviewCount;
     private final int likeCount;
